requirements.txt
fastapi==0.110.2
uvicorn[standard]==0.29.0
python-can==4.3.1
cantools==39.4.7
python-multipart==0.0.20
